Multiple bug fixes
1. deldomain.c: multiple bug fixes
2. deluser.c: skip vdelfiles when directory doesn't exist
3. limits.c: removed spurious warning message when limits doesn't exist
4. sql_renamedomain.c: refactored code
5. sql_setpw.c: bug fix - record not getting updated
6. vmoduser.c: removed incorrect call to vdeldomain
7. vmoveuser.c: bug fix - remove domain component from User
8. vrenamedomain.c: multiple bug fixes
9. vrenameuser.c: set BASE PATH of original domain for the new user
10. renameuser.c: ignore duplicate error when updating lastauth table

+++ indimail-x/vrenamedomain.c
@@ -1,5 +1,8 @@
 /*
  * $Log: vrenamedomain.c,v $
+ * Revision 1.6  2023-03-23 22:27:49+05:30  Cprogrammer
+ * multiple bug fixes
+ *
  * Revision 1.5  2023-03-20 10:39:05+05:30  Cprogrammer
  * standardize getln handling
  *
@@ -45,9 +48,10 @@
 #include <open.h>
 #include <getln.h>
 #include <env.h>
+#include <fmt.h>
 #include <replacestr.h>
+#include <setuserid.h>
 #endif
-#include "get_indimailuidgid.h"
 #include "variables.h"
 #include "check_group.h"
 #include "get_real_domain.h"
@@ -59,13 +63,14 @@
 #include "add_domain_assign.h"
 #include "del_domain_assign.h"
 #include "add_control.h"
+#include "del_control.h"
 #include "CreateDomainDirs.h"
 #include "is_alias_domain.h"
 #include "sql_getall.h"
 #include "post_handle.h"
 
 #ifndef	lint
-static char     sccsid[] = "$Id: vrenamedomain.c,v 1.5 2023-03-20 10:39:05+05:30 Cprogrammer Exp mbhangui $";
+static char     sccsid[] = "$Id: vrenamedomain.c,v 1.6 2023-03-23 22:27:49+05:30 Cprogrammer Exp mbhangui $";
 #endif
 
 #define WARN    "vrenamedomain: warning: "
@@ -74,8 +79,7 @@ static char     sccsid[] = "$Id: vrenamedomain.c,v 1.5 2023-03-20 10:39:05+05:30
 static void
 die_nomem()
 {
-	strerr_warn1("vrenamedomain: out of memory", 0);
-	_exit(111);
+	strerr_die2x(111, FATAL, "out of memory");
 }
 
 static int
@@ -84,68 +88,65 @@ getch(char *ch)
 	int             r;
 
 	if ((r = substdio_get(subfdinsmall, ch, 1)) == -1)
-		strerr_die1sys(111, "vrenamedomain: getch: unable to read input: ");
+		strerr_die2sys(111, FATAL, "getch: unable to read input: ");
 	return (r);
 }
 
 int
 main(int argc, char **argv)
 {
-	uid_t           uid;
-	gid_t           gid;
+	uid_t           uid, domainuid;
+	gid_t           gid, domaingid;
 	int             fd, i, match;
-	static stralloc OldDir = {0}, NewDir = {0}, tmpbuf = {0}, line = {0};
+	static stralloc OldDir = {0}, NewDir = {0}, TmpDir = {0},
+					tmpbuf = {0}, line = {0};
 	char           *real_domain, *ptr, *tmpstr, *base_argv0;
 	struct passwd  *pw;
 	struct stat     statbuf;
-	char            inbuf[4096], outbuf[512];
+	char            inbuf[4096], outbuf[512], linkbuf[512],
+					strnum1[FMT_ULONG], strnum2[FMT_ULONG];
 	struct substdio ssin, ssout;
 
 	if (argc != 3) {
 		strerr_warn1("USAGE: vrenamedomain old_domain_name new_domain_name", 0);
 		return (1);
 	}
-	if (indimailuid == -1 || indimailgid == -1)
-		get_indimailuidgid(&indimailuid, &indimailgid);
+	if (!get_assign(argv[1], &OldDir, &domainuid, &domaingid))
+		strerr_die3x(1, WARN, argv[1], ": domain does not exist");
+	if (!domainuid)
+		strerr_die4x(100, WARN, "domain ", argv[1], " with uid 0");
 	uid = getuid();
 	gid = getgid();
-	if (uid != 0 && uid != indimailuid && gid != indimailgid && check_group(indimailgid, FATAL) != 1)
-		strerr_die1x(100, "you must be root or indimail to run this program");
+	if (uid != 0 && uid != domainuid && gid != domaingid && check_group(domaingid, FATAL) != 1) {
+		strnum1[fmt_ulong(strnum1, domainuid)] = 0;
+		strnum2[fmt_ulong(strnum2, domaingid)] = 0;
+		strerr_die6x(100, WARN, "you must be root or domain user (uid=", strnum1, ", gid=", strnum2, ") to run this program");
+	}
 	if (uid && setuid(0))
-		strerr_die2sys(111, FATAL, "setuid: ");
-	if ((real_domain = get_real_domain(argv[2])) != (char *) 0) {
-		strerr_warn5("vrenamedomain: domain ", argv[2], " exists [", real_domain, "]", 0);
-		return (1);
-	} else
-	if (get_assign(argv[2], &tmpbuf, &uid, &gid)) {
-		strerr_warn2(argv[2], ": domain exists", 0);
-		return (1);
-	} else
+		strerr_die2sys(111, FATAL, "setuid-root: ");
+	if ((real_domain = get_real_domain(argv[2])) != (char *) 0)
+		strerr_die3x(1, WARN, argv[2], ": No such domain");
+	else
+	if (get_assign(argv[2], &tmpbuf, 0, 0))
+		strerr_die3x(1, WARN, argv[2], ": domain exists");
+	else
 	if (is_alias_domain(argv[1]) == 1) {
-		if (!(real_domain = get_real_domain(argv[1]))) {
-			strerr_warn2(argv[1], ": domain does not exist", 0);
-			return (1);
-		}
+		if (!(real_domain = get_real_domain(argv[1])))
+			strerr_die3x(1, WARN, argv[1], ": No such domain");
 		subprintfe(subfdout, "vrenamedomain", "Renaming alias domain %s (%s) to %s\n", argv[1], real_domain, argv[2]);
 		flush("vrenamedomain");
 		if (addaliasdomain(real_domain, argv[2]))
-			strerr_warn4("vaddaliasdomain: ", real_domain, " -> ", argv[2], 0);
+			strerr_warn4("addaliasdomain: ", real_domain, " -> ", argv[2], 0);
 		else
 		if (deldomain(argv[1]))
-			strerr_warn2("vdeldomain: ", argv[1], 0);
+			strerr_warn2("deldomain: ", argv[1], 0);
 		return (0);
-	} else
-	if (!get_assign(argv[1], &OldDir, &uid, &gid)) {
-		strerr_warn2(argv[1], ": failed to read assign file", 0);
-		return (1);
 	}
 	tmpbuf.len = 0;
 	if ((i = replacestr(OldDir.s, argv[1], argv[2], &tmpbuf)) == -1)
 		die_nomem();
-	if (i && rename(OldDir.s, tmpbuf.s)) {
-		strerr_warn4("vrenamedomain: rename: ", OldDir.s, " -> ", tmpbuf.s, &strerr_sys);
-		return (1);
-	}
+	if (i && rename(OldDir.s, tmpbuf.s))
+		strerr_die6sys(1, FATAL, "rename: ", OldDir.s, " -> ", tmpbuf.s, ": ");
 	if (!stralloc_copy(&NewDir, &tmpbuf) || !stralloc_0(&NewDir))
 		die_nomem();
 	NewDir.len--;
@@ -154,16 +155,16 @@ main(int argc, char **argv)
 	if ((tmpstr = str_str(tmpbuf.s, "/domains")) != (char *) 0)
 		*tmpstr = 0;
 	if (sql_renamedomain(argv[1], argv[2], NewDir.s))
-		return (1);
+		strerr_die6x(1, WARN, "sql_renamedomain: ", argv[1], " --> ", argv[2], " failed");
 	else
-	if (add_domain_assign(argv[2], tmpbuf.s, uid, gid))
-		return (1);
+	if (add_domain_assign(argv[2], tmpbuf.s, domainuid, domaingid))
+		strerr_die3x(1, WARN, "add_domain_assign: ", argv[2]);
 	else
 	if (add_control(argv[2], 0))
-		return (1);
+		strerr_die3x(1, WARN, "add_control: ", argv[2]);
 	else
-	if (del_domain_assign(argv[1], OldDir.s, uid, gid))
-		return (1);
+	if (del_domain_assign(argv[1], OldDir.s, domainuid, domaingid))
+		strerr_die3x(1, WARN, "add_domain_assign: ", argv[1]);
 	CreateDomainDirs(argv[2], uid, gid);
 	if (tmpstr && !*tmpstr) /*- restore the dir name in tmpbuf */
 		*tmpstr = '/';
@@ -178,9 +179,10 @@ main(int argc, char **argv)
 	if (fd != -1) {
 		substdio_fdbuf(&ssin, read, fd, inbuf, sizeof(inbuf));
 		subprintfe(subfdout, "vrenamedomain", "Relinking domains aliased to %s\n", argv[1]);
+		substdio_flush(subfdout);
 		for (;;) {
 			if (getln(&ssin, &line, &match, '\n') == -1) {
-				strerr_warn3("vrenamedomain: read: ", tmpbuf.s, ": ", &strerr_sys);
+				strerr_warn4(WARN, "read: ", tmpbuf.s, ": ", &strerr_sys);
 				close(fd);
 				return (1);
 			}
@@ -196,37 +198,51 @@ main(int argc, char **argv)
 					die_nomem();
 				line.len--;
 			}
-			if (!is_alias_domain(line.s)) {
-				strerr_warn2(line.s, ": Not an alias domain", 0);
+			if (!get_assign(line.s, &TmpDir, &uid, &gid)) {
+				strerr_warn4(WARN, "Domain ", line.s, " does not exist", 0);
 				continue;
 			} else
-			if (!get_assign(line.s, &OldDir, &uid, &gid)) {
-				strerr_warn3("Domain ", line.s, " does not exist", 0);
-				continue;
-			} else
-			if (lstat(OldDir.s, &statbuf)) {
-				strerr_warn2(OldDir.s, ": lstat: ", &strerr_sys);
+			if (lstat(TmpDir.s, &statbuf)) {
+				strerr_warn3(WARN, TmpDir.s, ": lstat: ", &strerr_sys);
 				continue;
 			} else
 			if (!S_ISLNK(statbuf.st_mode)) {
-				strerr_warn4(OldDir.s, " (", line.s, "): Not an alias domain", 0);
+				strerr_warn5(WARN, TmpDir.s, " (", line.s, "): Not an alias domain", 0);
+				continue;
+			}
+			if ((i = readlink(TmpDir.s, linkbuf, sizeof(linkbuf))) == -1) {
+				strerr_warn4(WARN, "readlink: ", TmpDir.s, ": ", &strerr_sys);
+				return -1;
+			}
+			if (i == sizeof(linkbuf)) {
+				errno = ENAMETOOLONG;
+				strerr_warn4(WARN, "readlink: ", TmpDir.s, ": ", &strerr_sys);
+				return -1;
+			}
+			linkbuf[i] = 0;
+			if (str_diffn(linkbuf, argv[1], i + 1))
+				continue; /*- this shouldn't happen */
+			/*-
+			 * we have already renamed OldDir
+			 * remove the dangling symlink and
+			 * relink to the renamed domain
+			 */
+			if (unlink(TmpDir.s)) {
+				strerr_warn3(WARN, TmpDir.s, ": unlink: ", &strerr_sys);
 				continue;
 			} else
-			if (unlink(OldDir.s)) {
-				strerr_warn2(OldDir.s, ": unlink: ", &strerr_sys);
-				continue;
-			}
-			if (symlink(NewDir.s, OldDir.s)) {
+			if (symlink(NewDir.s, TmpDir.s)) {
 				char            ch[1];
-				strerr_warn5("vrenamedomain: symlink: ", OldDir.s, " -> ", NewDir.s, ": ", &strerr_sys);
+				strerr_warn6(WARN, "symlink: ", TmpDir.s, " -> ", NewDir.s, ": ", &strerr_sys);
 				getch(ch);
 				continue;
+			} else {
+				subprintfe(subfdout, "vrenamedomain", "Linked Domain %s to %s [%s->%s]\n", line.s, argv[2], TmpDir.s, NewDir.s);
+				flush("vrenamedomain");
 			}
-			subprintfe(subfdout, "vrenamedomain", "Linked Domain %s to %s [%s->%s]\n", line.s, argv[2], OldDir.s, NewDir.s);
-			flush("vrenamedomain");
 		}
 	}
-	tmpbuf.len -= 14;
+	tmpbuf.len -= 14; /*- 14 = length of /.aliasdomains */
 	if (!stralloc_catb(&tmpbuf, "/.domain_rename", 15) || !stralloc_0(&tmpbuf))
 		die_nomem();
 	close(open(tmpbuf.s, O_TRUNC|O_CREAT, INDIMAIL_QMAIL_MODE));
@@ -245,25 +261,28 @@ main(int argc, char **argv)
 			die_nomem();
 		tmpbuf.len--;
 		if ((fd = open_trunc(tmpbuf.s)) == -1) {
-			strerr_warn3("vrenamedomain: open_trunc: ", tmpbuf.s, ": ", &strerr_sys);
+			strerr_warn4(WARN, "open_trunc: ", tmpbuf.s, ": ", &strerr_sys);
 			continue;
 		}
 		substdio_fdbuf(&ssout, write, fd, outbuf, sizeof(outbuf));
 		if (substdio_puts(&ssout, argv[1]) ||
 				substdio_put(&ssout, "\n", 1) ||
 				substdio_flush(&ssout)) {
-			strerr_warn3("vrenamedomain: write: ", tmpbuf.s, ": ", &strerr_sys);
+			strerr_warn4(WARN, "write: ", tmpbuf.s, ": ", &strerr_sys);
 			close(fd);
 			continue;
 		}
 		close(fd);
 	}
-	if (stralloc_copy(&tmpbuf, &NewDir) ||
+	/*- delete the old domain from the qmail control files */
+	if (del_control(argv[1]) == -1)
+		return (-1);
+	if (!stralloc_copy(&tmpbuf, &NewDir) ||
 			!stralloc_catb(&tmpbuf, "/.domain_rename", 15) ||
 			!stralloc_0(&tmpbuf))
 		die_nomem();
 	if (unlink(tmpbuf.s))
-		strerr_warn3("vrenamedomain: unlink: ", tmpbuf.s, ": ", &strerr_sys);
+		strerr_warn4(WARN, "unlink: ", tmpbuf.s, ": ", &strerr_sys);
 	if (!(ptr = env_get("POST_HANDLE"))) {
 		i = str_rchr(argv[0], '/');
 		if (!*(base_argv0 = (argv[0] + i)))
@@ -271,6 +290,12 @@ main(int argc, char **argv)
 		else
 			base_argv0++;
 		return(post_handle("%s/%s %s %s", LIBEXECDIR, base_argv0, argv[1], argv[2]));
-	} else
+	} else {
+		if (setuser_privileges(domainuid, domaingid, "indimail")) {
+			strnum1[fmt_ulong(strnum1, uid)] = 0;
+			strnum2[fmt_ulong(strnum2, gid)] = 0;
+			strerr_die6sys(111, FATAL, "setuser_privilege: (", strnum1, "/", strnum2, "): ");
+		}
 		return(post_handle("%s %s %s", ptr, argv[1], argv[2]));
+	}
 }
